Former-commit-id: 24d7e450fd1fa08817fc0bf98ea6d7b5a7df6658

+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Events/Interview/Base/QuestionAnswered.cs
@@ -6,7 +6,7 @@ namespace WB.Core.SharedKernels.DataCollection.Events.Interview.Base
     {
         public DateTime AnswerTime { get; private set; }
 
-        protected QuestionAnswered(Guid userId, Guid questionId, int[] propagationVector, DateTime answerTime)
+        protected QuestionAnswered(Guid userId, Guid questionId, decimal[] propagationVector, DateTime answerTime)
             : base(userId, questionId, propagationVector)
         {
             this.AnswerTime = answerTime;
